% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeckstats-package.R
\docType{package}
\name{adeckstats-package}
\alias{adeckstats}
\alias{adeckstats-package}
\title{adeckstats: Simulation of Distributions for Randomly Generated Values}
\description{
The package will help with proper understanding of statistics from a Theoretical programmer's perspective.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.dontbotheradeck.com}
  \item Report bugs at \url{https://www.dontbotheradeck.com}
}

}
\author{
\strong{Maintainer}: Tingwei Adeck \email{awesome.tingwei@outlook.com} (\href{https://orcid.org/0009-0009-7450-8863}{ORCID})

}
\keyword{internal}
