% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aalgorithms.R
\name{binarysearch}
\alias{binarysearch}
\alias{adeck_binarysearch}
\alias{.adeck_binarysearch}
\alias{adeck_binarysearch_mult}
\alias{adeck_binarysearch_recursive}
\title{Binary Search}
\usage{
adeck_binarysearch(vect, key = NULL, fun = NULL)

.adeck_binarysearch(vect, key = NULL, fun = NULL)

adeck_binarysearch_mult(vect, key = NULL, fun = NULL)

adeck_binarysearch_recursive(vect, key = NULL, fun = NULL)
}
\arguments{
\item{vect}{vector}

\item{key}{search value}

\item{fun}{boolean init}
}
\value{
index

index

index

index
}
\description{
Binary Search
}
\note{
The original function I wrote while learning the binary search algorithm
}
\concept{binarysearch}
