% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_broken_packages}
\alias{check_broken_packages}
\title{Check Broken Packages}
\usage{
check_broken_packages()
}
\description{
Check Broken Packages
}
\seealso{
Other utils: 
\code{\link{coins_mat_to_list}()},
\code{\link{combine_tibbles}()},
\code{\link{extract_rows_by_simnum}()},
\code{\link{list_of_vectors}()}
}
\concept{dev_helpers}
\concept{utils}
