% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coins_mat_to_list}
\alias{coins_mat_to_list}
\title{Coins Matrix to List}
\usage{
coins_mat_to_list(df, num_coins)
}
\arguments{
\item{df}{df}

\item{num_coins}{number of coins}
}
\value{
list
}
\description{
Coins Matrix to List
}
\seealso{
Other utils: 
\code{\link{check_broken_packages}()},
\code{\link{combine_tibbles}()},
\code{\link{extract_rows_by_simnum}()},
\code{\link{list_of_vectors}()}
}
\concept{permutation_helpers}
\concept{utils}
