% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_rows_by_simnum}
\alias{extract_rows_by_simnum}
\alias{extract_rows_by_mu_sigma}
\title{Extract tibbles}
\usage{
extract_rows_by_simnum(df)

extract_rows_by_mu_sigma(df)
}
\arguments{
\item{df}{df}
}
\value{
list

list
}
\description{
Extract tibbles

Extract tibbles
}
\seealso{
Other utils: 
\code{\link{check_broken_packages}()},
\code{\link{coins_mat_to_list}()},
\code{\link{combine_tibbles}()},
\code{\link{list_of_vectors}()}

Other utils: 
\code{\link{check_broken_packages}()},
\code{\link{coins_mat_to_list}()},
\code{\link{combine_tibbles}()},
\code{\link{list_of_vectors}()}
}
\concept{extractors}
\concept{utils}
