% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aalgorithms.R
\name{indexsearch}
\alias{indexsearch}
\alias{adeck_indexsearch}
\alias{adeck_indexsearch_mult}
\title{Index Search}
\usage{
adeck_indexsearch(vect, key = NULL, fun = NULL)

adeck_indexsearch_mult(vect, key = NULL, fun = NULL)
}
\arguments{
\item{vect}{vector}

\item{key}{search value}

\item{fun}{boolean init}
}
\value{
index
}
\description{
Index Search
}
\concept{indexsearch}
