% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_of_vectors}
\alias{list_of_vectors}
\title{List of Vectors}
\usage{
list_of_vectors(cp)
}
\arguments{
\item{cp}{Cartesian product}
}
\value{
lov
}
\description{
List of Vectors
}
\seealso{
Other utils: 
\code{\link{check_broken_packages}()},
\code{\link{coins_mat_to_list}()},
\code{\link{combine_tibbles}()},
\code{\link{extract_rows_by_simnum}()}
}
\concept{utils}
