% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anormsim.R
\name{normalsim}
\alias{normalsim}
\alias{adeck_normsim}
\alias{adeck_normsim.default}
\alias{.adeck_normsim}
\title{Normal Distribution Simulation}
\usage{
adeck_normsim(.n = NULL, .mu = NULL, .sd = NULL, .num_sims = NULL, rep = FALSE)

\method{adeck_normsim}{default}(.n = NULL, .mu = NULL, .sd = NULL, .num_sims = NULL, rep = FALSE)

.adeck_normsim(
  .n = NULL,
  .mu = NULL,
  .sd = NULL,
  .num_sims = NULL,
  rep = FALSE
)
}
\arguments{
\item{.n}{n}

\item{.mu}{mean}

\item{.sd}{sd}

\item{.num_sims}{sims}

\item{rep}{repeat}
}
\value{
list of tibbles

list of tibbles

list of tibbles
}
\description{
Normal Distribution Simulation
}
\examples{
\dontrun{
adeck_normsim(10,1,2)
}
}
\seealso{
\code{\link[adeckstats]{adeck_normsim_by_mu_sigma}}
}
\keyword{internal}
