% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anormsim.R
\name{normalsim_by_mu_sigma}
\alias{normalsim_by_mu_sigma}
\alias{adeck_normsim_by_mu_sigma}
\alias{adeck_normsim_by_mu_sigma.default}
\alias{.adeck_normsim_by_mu_sigma}
\title{Normal Simulation by Mean SD Combination}
\usage{
adeck_normsim_by_mu_sigma(
  .n = NULL,
  .mu = NULL,
  .sd = NULL,
  out_format = c("list", "tibble")
)

\method{adeck_normsim_by_mu_sigma}{default}(
  .n = NULL,
  .mu = NULL,
  .sd = NULL,
  out_format = c("list", "dataframe")
)

.adeck_normsim_by_mu_sigma(
  .n = NULL,
  .mu = NULL,
  .sd = NULL,
  out_format = c("list", "dataframe")
)
}
\arguments{
\item{.n}{n}

\item{.mu}{mean}

\item{.sd}{standard deviation}

\item{out_format}{output format (list preferable)}
}
\value{
list of tibbles

list of tibbles

list of tibbles
}
\description{
Normal Simulation by Mean SD Combination
}
\examples{
\dontrun{
adeck_normsim_by_mu_sigma(10,c(0,1),c(1,2))
}
}
\seealso{
\code{\link[adeckstats]{adeck_normsim}}
}
\keyword{internal}
