% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apermutations.R
\name{adeck_perm}
\alias{adeck_perm}
\alias{adeck_coinperm}
\title{Basic permutation}
\usage{
adeck_perm(n)

adeck_coinperm(number_of_coins, out_format = c("char", "num", "list"))
}
\arguments{
\item{n}{an integer}

\item{number_of_coins}{int}

\item{out_format}{format type}
}
\value{
a matrix or data frame

out_format type
}
\description{
Basic permutation

Coin Toss Permutations
}
\concept{permutation}
