% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plots}
\alias{plots}
\alias{plot.adeck_normsim}
\alias{plot.adeck_norm_tbl}
\title{Adeckstats Plots}
\usage{
\method{plot}{adeck_normsim}(
  x,
  which_sim = 1,
  whichplots = 1,
  histogram = 1 \%in\% whichplots,
  density = 2 \%in\% whichplots,
  ...
)

\method{plot}{adeck_norm_tbl}(
  x,
  whichplots = 1,
  histogram = 1 \%in\% whichplots,
  density = 2 \%in\% whichplots,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{plot requirement}

\item{which_sim}{sim number}

\item{whichplots}{int}

\item{histogram}{boolean}

\item{density}{boolean}

\item{...}{dots}

\item{xlab}{xlab}

\item{ylab}{ylab}
}
\value{
plot object

plot object

plot object
}
\description{
Adeckstats Plots
}
\concept{plots}
